backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
sqlalchemy==2.0.31
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow>=10.0.0
aiofiles==24.1.0
boto3>=1.34.0
mangum>=0.17.0

# --- Local AI / ML dependencies ---
transformers>=4.40.0
torch>=2.2.0
openai-whisper>=20240930
opencv-python>=4.9.0
pytrends>=4.9.2
requests>=2.31.0
moviepy>=1.0.3
